More Quartermaster alt in Loadout (#8942)
## About The Pull Request
A small little PR that brings some of the alt clothing for QM from the
Command vending machine to the Loadout
## Why It's Good For The Game
So the main reason why I created this PR is that I noticed only a few of
the alt quartermaster clothes were in the loadout (IE QM's Gorka and
QM's Skirtleneck) and the rest of the clothing choices were in the
Command vending machine that is out of QM's access. So I've included the
QM's Turtleneck and Casualwear to the loadout. I didn't bring over the
formal uniforms due to the rest of command not having theirs in the
loadout and I didnt want to give QM super special treatment.
## Changelog
:cl:
add: Adds Quartermaster Turtleneck and Casualwear to Loadout
/:cl: