The BSA now has a 5 second fire cooldown (#9285)
## About The Pull Request
Adds a 5 second delay (the same duration as the beam that shows up when
firing) to the BSA being allowed to fire again.
## Why It's Good For The Game
Some people have an issue with flashing lights. Firing the BSA at
lavaland 60 times in 10 seconds is very much a factor to cause this,
should help be better.
## Testing

I have GIFs demonstrating but they are too large for the PR. I'll post
them on discord.
## Changelog
:cl:
balance: The BSA can only fire once every 5 seconds.
/:cl:
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.