TYPO SPOTTED, NUKE THIS THING FROM ORBIT (#9304)
## About The Pull Request
Fixes a typo
## Why It's Good For The Game
i have a bomb strapped to my chest. it will explode in 5-10 business
years if this typo is not removed.
## Testing
Ummm... uhh... look! an eagle! *runs away from my problems*
## Changelog
:cl:
spellcheck: fixed a typo in the NV Meson cyborg upgrade.
spellcheck: updated the description of the NV Meson cyborg upgrade to
not use the advanced mop description.
/:cl:
## Pre-Merge Checklist
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.