Adds loafer spawers to maps that don't have them (#9223)
## About The Pull Request

this makes it so ALL maps have a
`/obj/structure/disposalconstruct/loafer` somewhere.

this doesn't guarantee a loafer spawn, as it's only a 20% chance if i
read the code correctly.

also moved the icebox one to a less stupid spot.

## Changelog
:cl:
map: All maps should have a chance to have a loafer somewhere in maints
now.
/:cl: