Alternative: Remove optical sensors from printing. (#8527)
## About The Pull Request
- Removes optical sensors from being printed.
- Closes: #8080 
## Why It's Good For The Game
As argued in #8080 Optical sensor are equal to shielded eyes in
strength, are a round start organ with very little downside, and can be
printed on mass. So the alternative suggestion to making it into a
welding goggles is just remove it's accessibility beyond being
roundstart.

Research options should not be equivalent to roundstart ones, that just
makes the science involved to get the former redundant.
## Changelog
:cl:
del: Optical sensors can no longer be printed.
/:cl: