Explorer gas mask protect against pepperspray like every other gas mask (#9581)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
- Explorer's gas masks now have pepper spray protection
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
- Because Ashdrake

So this was orginal reason it was removed on TG was as follows `this is
the strongest mask in the game and has no fov, it shouldn't also have
pepper spray protection. miners are not encountering pepper spray on
lavaland`
Now most of our masks do not FOV restrictions. Not that I'm familiar
with asides from the welding ones. But, we also now have a pepperspray
effect on lavaland that actually affects miners, Specially water. Trying
to put yourself out while on fire will cause confusion, most common and
prevalent with ash drake and the ice dragons I can't honestly remember
the name of. The only consistent way to remove this is with fire tank.
To which positioning is highly important with these two enemies. Since
we have unarguably refused the confusion effect time and time again. I
see it as only fair to reinstate the pepperproofing as there is a pepper
proofing effect that is critical as a miner, as not every miner with
anime quirk is a oozeling.
Any aspect petitioning the removal or adjustment of the confusion has
hit dead in the water (pun intended) so thus, bring back pepperspray
protection, lest a better alternative is found.

TLDR I want to enjoy having my small aesthetic change without dying to
RNG on ash drake.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Testing

<!-- Describe the tests you ran with your addition. It is recommended to
add images, videos and step-by-step explanations of conducted testing.
-->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
balance: Explorer masks have pepper spray protection
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->