transformation sting tweaks (#9781)
## About The Pull Request

This makes a couple of changes to transform sting:
- You now need 1 absorption in order to get it, just like with
teratomas.
- You can take mutadone in order to speed up how fast it wears off. It's
not instant, it just reduces the duration while in your system.

## Why It's Good For The Game

STOP SPAMMING THIS SHIT IT'S NOT FUNNY FOR THE MORBILLIONTH TIME. USE IT
STRATEGICALLY LIKE YOU'RE SUPPOSED TO.

## Testing
## Changelog
:cl:
balance: Changelings now need 1 absorb in order to obtain transformation
sting.
add: Mutadone can be used to heavily reduce the duration of
transformation sting while it's in your system.
/:cl:
## Pre-Merge Checklist
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.