Non-bloodsuckers/vassals can now resist out of silver cuffs while moving (#9680)
## About The Pull Request

this makes it so silver cuffs are less effective against non-vampires,
in a way that's harder to use as an easy antag test:
non-vampires/thralls will not have resisting canceled by being moved

draft bc i need to test it

## Why It's Good For The Game

another reason to ensure that silver cuffs are only used when necessary,
instead of potentially being used on everyone.

## Testing
## Changelog
:cl:
balance: When a non-bloodsucker/vassal is breaking out of silver cuffs,
the resist will not be canceled by moving them.
/:cl:
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.