Allow revived lavaland elites to speak (#9841)
## About The Pull Request

this makes it so reviving a lavaland elite with the tumro shard allows
them to talk.

also slightly cleaned up some code

Closes https://github.com/Monkestation/Monkestation2.0/issues/9823
Closes https://github.com/Monkestation/Monkestation2.0/issues/9428

## Why It's Good For The Game

this confuses people quite a bit, and allows for them to at least
attempt to RP.

## Changelog
:cl:
qol: Revived lavaland elites (as in, a sentient lavaland elite that is
revived after being defeated) can now speak.
/:cl: