[BALANCE] Reduces security assistant slots and updates their description, +1 minimum secoff slot to compensate (#9199)
## About The Pull Request

Discord discussion thread:
https://discord.com/channels/748354466335686736/1261773000013910029/threads/1431121926843273346

- Reduces the number of security assistant slots from 5 to 2.
- Increases the _minimum_ number of security officer slots from 5 to 6
to compensate. This can still be scaled higher based on server
population.
- Also updates the description of security assistant (and one of its
voucher kits) to not imply that it's a "learn to play sec" role.

## Why It's Good For The Game

Largely based off discussion in this discord thread:
https://discord.com/channels/748354466335686736/1430657119241113640

To summarize, security assistant currently struggles from a real
disconnect between its intended identity and the way it's often played
on the live game servers. Security assistants _very_ often tend to
simply become "officer lite" for a few simple reasons:
- There are too many slots for their intended "desk jockey/hall monitor"
purpose, so a lot of secasses end up with nothing better to do.
- They have a mindshield and antag rolling metaprotections, so the rest
of security generally has no problem with giving them extra gear.

Nothing better to do + a mindshield + easy access to extra gear =
security officer.

Reducing the slots was a no-brainer. For the other issues though, I
thought about removing the mindshield/metaprotections, but after
discussion decided that this was a way more funny solution to encourage
secasses to not loot the armory along with actual secoffs. You still CAN
pick up a gun and try to use it, but you're gonna be at a serious
disadvantage, so you are encouraged not to do so outside of true
emergencies.

Security assistant is also not good for learning to play security, never
has been good for learning to play security, and should not be
advertised as being good for learning to play security. To learn to play
security, you have to practice with and use actual security gear, not
shitty cucked versions of it. Of course you CAN simply ask for extra
gear as secass... but if you know what gear to ask for, you aren't new
to security anyways.

## Testing

Was tested on localhost - both roundstart and latejoin security
assistants get the trait + the warning in chat.

## Changelog
:cl:
balance: Security assistant slots reduced from 5 to 2.
balance: Minimum security officer slots increased from 5 to 6 (they can
also still be scaled higher with population).
/:cl:
## Pre-Merge Checklist
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.