medical circuits can no longer be installed in BCIs (#10146)
## About The Pull Request
title.

yes this is a balance change during a feature freeze. oversight maybe?
idk wirechecm was jammed into the codebase.

## Why It's Good For The Game
unfun to play with players with brain chips pumping them full of chems
and mixing pulses and lightning with no downside to themselves.

## Changelog

:cl:
balance: medical circuits cannot be installed in brain computer
interfaces.
/:cl:

## Pre-Merge Checklist

- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.

---------

Co-authored-by: Lucy <lucy@absolucy.moe>